backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.0.1
rembg==2.0.50
opencv-python==4.8.1.78
numpy==1.24.3
mediapipe==0.10.14
python-dotenv==1.0.0